Change all double to Double, add Integrator plant object

+++ PID/src/objects/mathfunctions/Integral.java
@@ -2,8 +2,10 @@ package objects.mathfunctions;
 
 import objects.GlobalSettings;
 
+import java.util.zip.DeflaterOutputStream;
+
 public final class Integral implements GlobalSettings {
-    private double Ts;                 //sample time in milliseconds
+    private Double Ts;                 //sample time in milliseconds
     private double integralVal;
     
     public Integral(double sampleTime, double initialCondition) {
@@ -19,7 +21,7 @@ public final class Integral implements GlobalSettings {
         integralVal = 0;
     }
     
-    public void resetIntegral(int newSampleTime) {
+    public void resetIntegral(Double newSampleTime) {
         this.Ts = newSampleTime;
         integralVal = 0;
     }
@@ -41,7 +43,7 @@ public final class Integral implements GlobalSettings {
     
     //Interface
     @Override
-    public void setSampleTime(double Ts) {
+    public void setSampleTime(Double Ts) {
         this.Ts = Ts;
     }
 }
